LayoutTests/imported/w3c/web-platform-tests/css/css-font-loading/fontface-loadingevent.html

<!DOCTYPE html>
<title>Tests FontFaceSet loading event</title>
<link rel="help" href="https://drafts.csswg.org/css-font-loading/#events">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async t => {
  const fontSet = document.fonts;
  let loadingFired = false;

  fontSet.addEventListener("loading", (e) => {
    assert_not_equals(e.fontfaces.findIndex(fontface => fontface.family === "TestFont1"), -1, "The 'loading' event.fontfaces contained `TestFont1`");
    assert_not_equals(e.fontfaces.findIndex(fontface => fontface.family === "TestFont2"), -1, "The 'loading' event.fontfaces contained `TestFont2`");
    loadingFired = true;
  });
  
  const testFont1 = new FontFace("TestFont1", "url(/fonts/Ahem.ttf)");
  const testFont2 = new FontFace("TestFont2", "url(/fonts/Ahem.ttf)");
  fontSet.add(testFont1);
  fontSet.add(testFont2);

  await Promise.all([testFont1.load(), testFont2.load()]);

  await fontSet.ready;
  assert_true(loadingFired, "The 'loading' event should have fired");
}, "FontFaceSet fires correct loading event, with the loading fontfaces");
</script>
